requirements.txt
apache-airflow==2.9.1
apache-airflow-providers-postgres==5.11.1
pandas==2.2.2
psycopg2-binary==2.9.9
requests==2.31.0
jupyter==1.0.0
